modules/uppic/backend/requirements.txt
# Uppic 独立微服务依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
boto3==1.35.0
python-dotenv==1.0.0
